requirements-dev.txt
# Ludax in development mode
-e .[gui, agents]

# Graphs
matplotlib
pandas
thefuzz
tqdm

# Alphazero
pgx>=2.0.0
dm-haiku
mctx
optax
wandb
omegaconf
pydantic
